Drop support from anon functions, add executed_steps
Now the only way to define a pipeline is by using modules, because
to use anonymous functions you would have to manipulate the state
manually, which is not intended.

The Pipeline.State struct now tracks which steps were executed as
well, usefull for debugging and for metadata collection.

it was not